backend/requirements.txt
Flask
Wireviz
PyYAML
